requirements.txt
openai>=1.30.0
mcp>=1.0.0
httpx>=0.27.0
rich>=13.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
